ad_pipeline/requirements.txt
streamlit
pillow
numpy
opencv-python
huggingface_hub
requests
